+++ tests/integ/test_df_to_pandas.py
@@ -51,7 +51,6 @@ from snowflake.snowpark.types import (
     TimeType,
     VariantType,
 )
-import snowflake.snowpark.session
 from tests.utils import IS_IN_STORED_PROC, Utils
 
 
@@ -450,27 +449,6 @@ def test_write_pandas_chunk_size(session, monkeypatch):
             "snowflake.snowpark.session.write_pandas", side_effect=write_pandas_wrapper
         ) as mock_write_pandas:
             session.create_dataframe(table, chunk_size=10)
-            # Verify that write_arrow was called once
-            mock_write_pandas.assert_called_once()
-
-        # Approach 2: Module variable update
-        original_pandas_chunk_size = snowflake.snowpark.session.WRITE_PANDAS_CHUNK_SIZE
-        assert (
-            original_pandas_chunk_size == 100000
-            if IS_IN_STORED_PROC
-            else original_pandas_chunk_size is None
-        )
-        monkeypatch.setattr(snowflake.snowpark.session, "WRITE_PANDAS_CHUNK_SIZE", 10)
-        assert snowflake.snowpark.session.WRITE_PANDAS_CHUNK_SIZE == 10
-
-        with mock.patch(
-            "snowflake.snowpark.session.write_pandas", side_effect=write_pandas_wrapper
-        ) as mock_write_pandas:
-            session.write_pandas(
-                table,
-                table_name,
-                auto_create_table=True,
-            )
             # Verify that write_arrow was called once
             mock_write_pandas.assert_called_once()
     finally: